Switch www container to host-mounted source (#4)
Added a new tag :live for the site container that mounts the source instead of building it into the container and updated the docker/www.sh script to launch that instead of :latest. This new container will be suitable for development

The build script is now way less verbose and dumps logs in the event that the build failed so it we lose no ability to debug.

Also ran dos2unix on README.md to unfuck line endings.